+++ astroquery/exoplanet_archive/data/exoplanet_nexsci_units.json
