+++ astroquery/exoplanets_org/__init__.py
@@ -2,6 +2,6 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-__all__ = ['PlanetParams']
+__all__ = ['ExoplanetsOrg']
 
-from .planet_params import PlanetParams
+from .exoplanets_org import ExoplanetsOrg
